+++ Changelog.md
@@ -2,5 +2,12 @@
 All notable changes will be documented in this file
 
 ## [Unreleased]
-- Create class called roboMethods, inheriting a class driveMethods
-- Implement driving algorithm as a function of driveMethods
+- Create a user interface to allow selection between multiple modes
+- A competition control system for driver control
+
+## C-MVEXO [0.0.0] - Planned Release: 2018-9-18
+### Added
+- Basic driver control
+- x-drive chassis driving
+- h-drive chassis driving
+- General PID calculator